refactor(FE): 사용자 설정 및 프로필 관리 기능을 ProfileButton 컴포넌트로 리팩토링
- UserSetting 컴포넌트를 제거하고, 해당 기능을 ProfileButton에 통합하여 모듈성을 향상시킴
- 테마 전환, 프로필 편집, 피드백 제출, 로그아웃, 계정 삭제를 포함한 사용자 프로필 작업을 위한 드롭다운 메뉴 추가
- 프로필 페이지에서 사용자 관련 작업을 위해 UserSetting 대신 ProfileButton을 사용하도록 업데이트
