+++ front/src/widgets/Header.tsx
@@ -1,6 +1,5 @@
 import Logo from "@/features/common/Logo";
 import NotificationButton from "@/widgets/NotificationButton";
-import KakaoLoginButton from "@/features/auth/KakaoLoginButton";
 import { useSelector } from "react-redux";
 import { RootState } from "@/state/store/index";
 import { Link } from "react-router-dom";
@@ -66,17 +65,14 @@ export default function Header() {
               </Link>
               <div className="flex flex-1 md:justify-end items-center gap-2">
                 <SearchBar />
-                {userInfo.kakaoId ? (
+                {userInfo.kakaoId !== 0 && (
                   // 로그인을 했을 경우 유저가 사용 가능한 버튼 보여주기
                   <>
                     <WriteReviewButton />
                     <NotificationButton />
-                    <ProfileButton userInfo={userInfo} />
                   </>
-                ) : (
-                  // 닉네임이 없으면 로그인 버튼 보여주기
-                  <KakaoLoginButton />
                 )}
+                <ProfileButton userInfo={userInfo} />
               </div>
             </div>
           </header>
